epdc_v2: hack-fix dma address printing

+++ drivers/video/fbdev/mxc/mxc_epdc_v2_fb.c
@@ -1249,7 +1249,7 @@ static void dump_collision_list(struct mxc_epdc_fb_data *fb_data)
 		dev_info(fb_data->dev, "Empty");
 	list_for_each_entry(plist, &fb_data->upd_buf_collision_list, list) {
 		dev_info(fb_data->dev, "Virt Addr = 0x%x, Phys Addr = 0x%x ",
-			(u32)plist->virt_addr, plist->phys_addr);
+			(u32)plist->virt_addr, (u32)plist->phys_addr);
 		dump_update_data(fb_data->dev, plist);
 	}
 }
@@ -1263,7 +1263,7 @@ static void dump_free_list(struct mxc_epdc_fb_data *fb_data)
 		dev_info(fb_data->dev, "Empty");
 	list_for_each_entry(plist, &fb_data->upd_buf_free_list, list)
 		dev_info(fb_data->dev, "Virt Addr = 0x%x, Phys Addr = 0x%x ",
-			(u32)plist->virt_addr, plist->phys_addr);
+			(u32)plist->virt_addr, (u32)plist->phys_addr);
 }
 
 static void dump_queue(struct mxc_epdc_fb_data *fb_data)
@@ -1275,7 +1275,7 @@ static void dump_queue(struct mxc_epdc_fb_data *fb_data)
 		dev_info(fb_data->dev, "Empty");
 	list_for_each_entry(plist, &fb_data->upd_buf_queue, list) {
 		dev_info(fb_data->dev, "Virt Addr = 0x%x, Phys Addr = 0x%x ",
-			(u32)plist->virt_addr, plist->phys_addr);
+			(u32)plist->virt_addr, (u32)plist->phys_addr);
 		dump_update_data(fb_data->dev, plist);
 	}
 }
